MINOR: Preserve the base offset of the original record batch in V2
The previous code did not handle this correctly if a batch was
compacted more than once.

Also add test case for duplicate check after log cleaning and
improve various comments.

Author: Jason Gustafson <jason@confluent.io>

Reviewers: Ismael Juma <ismael@juma.me.uk>

Closes #3145 from hachikuji/minor-improve-base-sequence-docs

+++ clients/src/main/java/org/apache/kafka/common/record/AbstractLegacyRecordBatch.java
@@ -186,7 +186,7 @@ public abstract class AbstractLegacyRecordBatch extends AbstractRecordBatch impl
     }
 
     @Override
-    public long sequence() {
+    public int sequence() {
         return RecordBatch.NO_SEQUENCE;
     }
